ver 0.6

+++ cse6730_project3/src/Parameter.java
@@ -12,9 +12,9 @@ public class Parameter {
 
     public static double antibody_ratio;
 
-    public static int max_vaccine;
-    public static int max_patient;
-    public static int max_quarantine;
+    public static double max_vaccine;
+    public static double max_patient;
+    public static double max_quarantine;
     public static double cure_rate;
 
     public static int threshold = 20;
@@ -23,7 +23,7 @@ public class Parameter {
     public static int max_air = 60, max_train = 40, max_ship = 20;
     public static double pop_scale = 3000.0;
 
-    public Parameter(int apm, int tpm, int spm, double asm, double tsm, double ssm, double rs, double rm, double rl, double ar, int mv, int mp, int mq, double cr) {
+    public Parameter(int apm, int tpm, int spm, double asm, double tsm, double ssm, double rs, double rm, double rl, double ar, double mv, double mp, double mq, double cr) {
         air_passenger_mean = apm;
         train_passenger_mean = tpm;
         ship_passenger_mean = spm;
